==> analysis/scripts/01_extract.Rmd <==
---
title: "R Notebook"
output: html_notebook
---


```{r}
library(tidyverse)
```

```{r}
files <- list.files('../../data/extracted', full.names = TRUE) %>% set_names(basename(.) %>% str_remove('.csv'))
files
```


```{r}
column_names <- c('Specialty',
                  'No. of Programs',
                  'Positions Offered',
                  'Unfilled Programs',
                  'U.S. Senior Applicants',
                  'Total Applicants',
                  'U.S. Senior Matches',
                  'Total Matches',
                  'U.S. Senior % Filled',
                  'Total % Filled',
                  'U.S. Senior Ranked Positions',
                  'Total Ranked Positions')

merged <- map(files, function(i) {
  dataset <- names(files)[files == i] %>% str_split(' - ') %>% unlist()
  
  read_csv(i, col_names = column_names) %>% 
    filter(!str_starts(Specialty, 'Total|Specialty')) %>% 
    drop_na() %>% 
    mutate(across(everything(), str_remove, pattern = ',')) %>% 
    mutate(across(!Specialty, as.numeric)) %>% 
    mutate(Year = dataset[1],
           Class = dataset[2])
    
}) %>% 
  bind_rows() %>% 
  select(Specialty, Class, Year, everything())

## Fix certain Specialty names
merged <- merged %>% 
  mutate(Specialty = case_when(
    Specialty == 'Psychiatry' & Class == 'PGY1' ~ 'Psychiatry (Categorical)',
    Specialty == 'Pediatrics-Prelim (PGY-1 Only)' ~ 'Pediatrics-Preliminary',
    TRUE ~ Specialty
  ))

merged
```

```{r}
merged %>% 
  group_by(Year, Class) %>% 
  tally() %>% 
  pivot_wider(names_from = Class,
              values_from = n)
```


```{r}
merged %>% 
  filter(Specialty == 'Dermatology') %>% arrange(Year) 
  
```

## Export
```{r}
merged %>% write_tsv('../output/merged_table.tsv')
```

## Session info
```{r}
sessionInfo()
```
